Finished data processing, switched to minmax normalization.
To use the class, simply create an ECGRecord object and pass in the record number. Then, convert it to an ECGData object. Now, there are two batches generated, one for individual beats and one for joined beats. Call the corresponding 'k-fold validation' function, and it will generate an xtrain, xtest, ytrain and ytest for you.